requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
numpy==2.2.1
opencv-python==4.10.0.84
Pillow==11.0.0
psd-tools==1.10.7
